Cleanup IPlugOSC classes, adding lambda for receiver
